Moved WishSender and dependency on okhttp to client library
