+++ app/build.gradle
@@ -92,7 +92,6 @@ dependencies {
     compile "com.android.support:support-v4:$support_v4_version"
     compile "com.fasterxml.jackson.core:jackson-databind:2.8.6"
     compile "org.springframework.android:spring-android-rest-template:1.0.1.RELEASE"
-    compile "com.squareup.okhttp3:okhttp:3.6.0"
     compile "org.slf4j:slf4j-android:1.7.22"
     compile "org.androidannotations:androidannotations-api:$androidannotations_version"
     compile "org.androidannotations:rest-spring-api:$androidannotations_version"
